moved admin_command method into main DB class
