+++ MongoBackup/DB.py
@@ -1,16 +1,19 @@
 import logging
 
 from pymongo import MongoClient
+from time import sleep
 
 
 class DB:
-    def __init__(self, host='localhost', port=27017, username=None, password=None, authdb="admin", conn_timeout=5000):
-        self.host = host
-        self.port = port
-        self.username = username
-        self.password = password
-        self.authdb = authdb
+    def __init__(self, host='localhost', port=27017, username=None, password=None, authdb="admin", conn_timeout=5000, retries=5):
+        self.host         = host
+        self.port         = port
+        self.username     = username
+        self.password     = password
+        self.authdb       = authdb
         self.conn_timeout = conn_timeout
+        self.retries      = retries
+
         self._conn = None
         self.connect()
         self.auth_if_required()
@@ -40,6 +43,22 @@ class DB:
                 raise e
         else:
             pass
+
+    def admin_command(self, admin_command):
+        tries  = 0
+        status = None
+        while not status and tries < self.retries:
+            try:
+                status = self._conn['admin'].command(admin_command)
+                if not status:
+                    raise e
+            except Exception, e:
+                logging.error("Error running admin command '%s': %s" % (admin_command, e))
+                tries += 1
+                sleep(1)
+        if not status:
+            raise Exception, "Could not get output from command: '%s' after %i retries!" % (admin_command, self.retries), None
+        return status
 
     def connection(self):
         return self._conn
